requirements.txt
streamlit
pandas
numpy
tensorflow-cpu
altair
scikit-learn